Create Ethereum Gas Tracker
Here is a pure function written in Solidity (a programming language for the Ethereum blockchain) that fetches the current gas price from the Ethereum Gas Tracker:

This function uses the getGasPrice() function from the Ethereum Gas Tracker contract, which returns a tuple that contains a boolean indicating whether the call was successful and the current gas price in wei. The function first checks if the call was successful using the require() function, and if it was, it returns the gas price.

You can call this function by calling getGasPrice() in a smart contract or by calling it via a web3.js function in a dapp.

Please note:

This function was written based on the solidity version 0.8.3, so you may want to double check the version you use.
You need to have a web3 provider in order to use this function.